# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MediCore Hospital Management SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      parameters:
        - name: doctorId
          in: query
          schema:
            type: integer
        - name: patientId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: departmentId
          in: query
          schema:
            type: integer
        - name: appointmentType
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Appointments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentList"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Cancel appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Cancelled
  /appointments/{id}/status:
    patch:
      operationId: updateAppointmentStatus
      tags: [appointments]
      summary: Update appointment status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentStatusInput"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/calendar:
    get:
      operationId: getCalendarAppointments
      tags: [appointments]
      summary: Get appointments for calendar view
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
        - name: endDate
          in: query
          required: true
          schema:
            type: string
        - name: doctorId
          in: query
          schema:
            type: integer
        - name: departmentId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Calendar appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarAppointment"
  /appointments/available-slots:
    get:
      operationId: getAvailableSlots
      tags: [appointments]
      summary: Get available time slots for booking
      parameters:
        - name: doctorId
          in: query
          required: true
          schema:
            type: integer
        - name: date
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Available slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeSlot"

  # ─── INVOICES ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Appointment:
      type: object
      required: [id, appointmentDate, status, tokenNumber]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        patientName:
          type: string
        patientPhone:
          type: ["string", "null"]
        doctorId:
          type: integer
        doctorName:
          type: string
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        appointmentDate:
          type: string
        appointmentTime:
          type: string
        tokenNumber:
          type: integer
        status:
          type: string
        appointmentType:
          type: string
        symptoms:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        roomNumber:
          type: ["string", "null"]
        paymentStatus:
          type: ["string", "null"]
        createdAt:
          type: string
    AppointmentList:
      type: object
      required: [appointments, total]
      properties:
        appointments:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    AppointmentInput:
      type: object
      required: [patientId, doctorId, appointmentDate, appointmentTime]
      properties:
        patientId:
          type: integer
        doctorId:
          type: integer
        departmentId:
          type: integer
        appointmentDate:
          type: string
        appointmentTime:
          type: string
        appointmentType:
          type: string
        symptoms:
          type: string
        notes:
          type: string
        roomNumber:
          type: string
    AppointmentUpdate:
      type: object
      properties:
        appointmentDate:
          type: string
        appointmentTime:
          type: string
        status:
          type: string
        appointmentType:
          type: string
        symptoms:
          type: string
        diagnosis:
          type: string
        notes:
          type: string
        followUpDate:
          type: ["string", "null"]
        roomNumber:
          type: string
    AppointmentStatusInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
        notes:
          type: ["string", "null"]
    CalendarAppointment:
      type: object
      required: [id, title, start, end]
      properties:
        id:
          type: integer
        title:
          type: string
        start:
          type: string
        end:
          type: string
        patientName:
          type: string
        doctorName:
          type: string
        departmentName:
          type: ["string", "null"]
        tokenNumber:
          type: integer
        status:
          type: string
        appointmentType:
          type: string
        paymentStatus:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
    TimeSlot:
      type: object
      required: [time, available]
      properties:
        time:
          type: string
        available:
          type: boolean
        tokenNumber:
          type: ["integer", "null"]

    # ─── INVOICE ──────────────────────────────────────────────────────────────
...